# openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Customer Support Q&A API",
    "version": "2.1.0",
    "description": "Authenticated API for policy-grounded customer-support answers."
  },
  "servers": [{"url": "http://127.0.0.1:3000"}],
  "paths": {
    "/health/live": {
      "get": {
        "summary": "Process liveness",
        "responses": {
          "200": {
            "description": "Process is running",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Health"}}}
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "summary": "Configuration readiness",
        "responses": {
          "200": {
            "description": "Required configuration is present",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ReadinessReady"}}}
          },
          "503": {
            "description": "Required configuration is missing",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ReadinessNotReady"}}}
          }
        }
      }
    },
    "/support/ask": {
      "post": {
        "summary": "Answer one customer-support question",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/QuestionRequest"}}}
        },
        "responses": {
          "200": {
            "description": "Completed provider answer",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AnswerResponse"}}}
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "413": {"$ref": "#/components/responses/PayloadTooLarge"},
          "429": {"$ref": "#/components/responses/RateLimited"},
          "500": {"$ref": "#/components/responses/InternalError"},
          "502": {"$ref": "#/components/responses/ProviderError"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"},
          "504": {"$ref": "#/components/responses/ProviderTimeout"}
        }
      }
    }
  },
  "components": {
    "securitySchemes": {"bearerAuth": {"type": "http", "scheme": "bearer"}},
    "headers": {
      "WWWAuthenticate": {"description": "Required authentication scheme", "schema": {"type": "string"}},
      "RetryAfter": {"description": "Seconds before retrying", "schema": {"type": "string"}},
      "RateLimit": {"description": "Current rate-limit state", "schema": {"type": "string"}},
      "RateLimitPolicy": {"description": "Applied rate-limit policy", "schema": {"type": "string"}}
    },
    "schemas": {
      "Health": {
        "type": "object",
        "additionalProperties": false,
        "required": ["status"],
        "properties": {"status": {"type": "string", "enum": ["ok"]}}
      },
      "ReadinessReady": {
        "type": "object",
        "additionalProperties": false,
        "required": ["status"],
        "properties": {"status": {"type": "string", "enum": ["ready"]}}
      },
      "ReadinessNotReady": {
        "type": "object",
        "additionalProperties": false,
        "required": ["status", "missing"],
        "properties": {
          "status": {"type": "string", "enum": ["not_ready"]},
          "missing": {
            "type": "array",
            "minItems": 1,
            "items": {"type": "string", "enum": ["OPENAI_API_KEY", "SUPPORT_API_KEY"]}
          }
        }
      },
      "QuestionRequest": {
        "type": "object",
        "additionalProperties": true,
        "required": ["question"],
        "properties": {"question": {"type": "string", "minLength": 1, "maxLength": 2000}}
      },
      "Usage": {
        "type": ["object", "null"],
        "properties": {
          "input_tokens": {"type": ["integer", "null"], "minimum": 0},
          "output_tokens": {"type": ["integer", "null"], "minimum": 0},
          "total_tokens": {"type": ["integer", "null"], "minimum": 0}
        }
      },
      "AnswerResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["answer", "model", "usage"],
        "properties": {
          "answer": {"type": "string", "minLength": 1},
          "model": {"type": "string", "minLength": 1},
          "usage": {"$ref": "#/components/schemas/Usage"}
        }
      },
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error", "message"],
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "auth_rate_limited",
              "internal_error",
              "invalid_json",
              "invalid_request",
              "not_found",
              "payload_too_large",
              "provider_error",
              "provider_rate_limited",
              "provider_timeout",
              "rate_limited",
              "sensitive_content",
              "service_not_ready",
              "unauthorized"
            ]
          },
          "message": {"type": "string", "minLength": 1}
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid JSON, request shape, or sensitive content",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}
      },
      "Unauthorized": {
        "description": "Missing or invalid bearer token",
        "headers": {"WWW-Authenticate": {"$ref": "#/components/headers/WWWAuthenticate"}},
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}
      },
      "PayloadTooLarge": {
        "description": "Request body exceeds 32 KB",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}
      },
      "RateLimited": {
        "description": "Authentication-failure or authenticated paid-request limit reached",
        "headers": {
          "Retry-After": {"$ref": "#/components/headers/RetryAfter"},
          "RateLimit": {"$ref": "#/components/headers/RateLimit"},
          "RateLimit-Policy": {"$ref": "#/components/headers/RateLimitPolicy"}
        },
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}
      },
      "InternalError": {
        "description": "Unexpected internal failure",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}
      },
      "ProviderError": {
        "description": "Provider did not return a complete answer",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}
      },
      "ServiceUnavailable": {
        "description": "Configuration or provider capacity unavailable",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}
      },
      "ProviderTimeout": {
        "description": "Provider timed out",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}
      }
    }
  }
}